+++ heuristic/sokoban_heuristic.py
@@ -28,4 +28,4 @@ class SokobanHeuristic(Heuristic):
         )
 
         # The heuristic value is the boxes missing from their correct positions
-        return (target_box_count - matching_boxes) * 5
+        return ((target_box_count - matching_boxes) * 5).astype(jnp.float32)
